requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
pillow==10.2.0
torch>=2.4.0
transformers>=4.46.1
unsloth>=2024.11.2
accelerate>=0.21.0
bitsandbytes>=0.41.0
datasets>=2.14.0
tqdm>=4.65.0
wandb>=0.15.0
trl>=0.7.4
safetensors>=0.4.1 